add back delete alert for deleting items

+++ Components/useHeaderNavigation.js
@@ -1,4 +1,5 @@
 import { useLayoutEffect } from 'react';
+import { Alert } from 'react-native';
 import Button from './Button';
 import { color, spacing, fontSize } from './StyleHelper';
 import { FontAwesome } from '@expo/vector-icons';
@@ -24,8 +25,24 @@ const useHeaderNavigation = (navigation, route, type) => {
    */
   const handleDelete = () => {
     console.log('Delete activity');
-    FirestoreService.deleteActivity(route.params.activity.id);
-    navigation.navigate('All Activities');
+    Alert.alert(
+      "Delete", 
+      "Are you sure you want to delete this item?",
+      [
+          {
+              text: "No",
+              onPress: () => console.log("Cancel Pressed"),
+              style: "cancel"
+          },
+          { 
+              text: "Yes", onPress: () => {
+                  FirestoreService.deleteActivity(route.params.activity.id);
+                  navigation.navigate('All Activities');
+              } 
+          }
+      ],
+      { cancelable: false }
+    );
   }
 
   // Set header navigation options based on the type of navigation.
